Reimplement remaining DatabaseMetaData methods and fix precision of binary numeric types

+++ h2/src/test/org/h2/test/scripts/dml/show.sql
@@ -77,17 +77,17 @@ SHOW TABLES FROM SCH;
 SHOW COLUMNS FROM TEST_P;
 > FIELD TYPE         NULL KEY DEFAULT
 > ----- ------------ ---- --- -------
-> ID_P  INTEGER(10)  NO   PRI NULL
+> ID_P  INT          NO   PRI NULL
 > U_P   VARCHAR(255) YES  UNI NULL
-> N_P   INTEGER(10)  YES      1
+> N_P   INT          YES      1
 > rows (ordered): 3
 
 SHOW COLUMNS FROM TEST_S FROM SCH;
 > FIELD TYPE         NULL KEY DEFAULT
 > ----- ------------ ---- --- -------
-> ID_S  INTEGER(10)  NO   PRI NULL
+> ID_S  INT          NO   PRI NULL
 > U_S   VARCHAR(255) YES  UNI NULL
-> N_S   INTEGER(10)  YES      1
+> N_S   INT          YES      1
 > rows (ordered): 3
 
 SHOW DATABASES;
